[nitro-protocol] Do not cast enum type to uint8

+++ packages/nitro-protocol/contracts/AssetHolder.sol
@@ -254,7 +254,7 @@ contract AssetHolder is IAssetHolder {
             assetOutcomeHashes[fromChannelId] = keccak256(
                 abi.encode(
                     Outcome.AssetOutcome(
-                        uint8(Outcome.AssetOutcomeType.Allocation),
+                        Outcome.AssetOutcomeType.Allocation,
                         abi.encode(newAllocation)
                     )
                 )
@@ -319,7 +319,7 @@ contract AssetHolder is IAssetHolder {
             assetOutcomeHashes[guarantee.targetChannelId] = keccak256(
                 abi.encode(
                     Outcome.AssetOutcome(
-                        uint8(Outcome.AssetOutcomeType.Allocation),
+                        Outcome.AssetOutcomeType.Allocation,
                         abi.encode(newAllocation)
                     )
                 )
@@ -405,10 +405,7 @@ contract AssetHolder is IAssetHolder {
             assetOutcomeHashes[channelId] ==
                 keccak256(
                     abi.encode(
-                        Outcome.AssetOutcome(
-                            uint8(Outcome.AssetOutcomeType.Allocation),
-                            allocationBytes
-                        )
+                        Outcome.AssetOutcome(Outcome.AssetOutcomeType.Allocation, allocationBytes)
                     )
                 ),
             'h(allocation)!=assetOutcomeHash'
@@ -423,10 +420,7 @@ contract AssetHolder is IAssetHolder {
             assetOutcomeHashes[guarantorChannelId] ==
                 keccak256(
                     abi.encode(
-                        Outcome.AssetOutcome(
-                            uint8(Outcome.AssetOutcomeType.Guarantee),
-                            guaranteeBytes
-                        )
+                        Outcome.AssetOutcome(Outcome.AssetOutcomeType.Guarantee, guaranteeBytes)
                     )
                 ),
             'h(guarantee)!=assetOutcomeHash'
